refactor: Add functionality to get pagination results - #41
